api/requirements.txt
asgiref==3.5.0
azure-common==1.1.28
azure-core==1.22.1
azure-identity==1.7.1
azure-keyvault-secrets==4.3.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
google-api-core==2.5.0
google-auth==2.6.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
idna==3.3
isodate==0.6.1
msal==1.16.0
msal-extensions==0.3.1
msrest==0.6.21
mssql-django==1.1.2
oauthlib==3.2.0
opencensus==0.8.0
opencensus-context==0.1.2
opencensus-ext-azure==1.1.1
opencensus-ext-django==0.7.6
opencensus-ext-ocagent==0.7.1
opencensus-proto==0.1.0
portalocker==2.3.2
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyodbc==4.0.32
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
